feat(workflows): add support for windows-arm64 in multiple build configurations and enhance slim binary handling
