+++ .github/workflows/manual-build-target.yml
@@ -9,6 +9,7 @@ on:
         options:
           - windows
           - windows7
+          - windows-arm64
           - darwin-amd64 
           - darwin-arm64
           - linux-amd64
@@ -21,6 +22,7 @@ on:
           - full
           - yakit
           - irify
+          - yakslim
         default: full
       ee:
         type: boolean
