+++ common/consts/yakit.go
@@ -1,6 +1,8 @@
 package consts
 
 import (
+	"os"
+	"path/filepath"
 	"sync"
 
 	"github.com/yaklang/yaklang/common/schema"
@@ -147,6 +149,14 @@ func initializeYakitDirectories() {
 	GetDefaultYakitBaseTempDir() // yakit-projects/temp
 	GetDefaultAISkillsDir()      // yakit-projects/ai-skills
 	GetDefaultYakitOpenAPIDocumentsDir() // yakit-projects/openapi-documents
+
+	utils.RegisterTempFileOpener(func(name string) (*os.File, error) {
+		dir := GetDefaultYakitBaseTempDir()
+		if !utils.IsDir(dir) {
+			_ = os.MkdirAll(dir, 0o755)
+		}
+		return os.OpenFile(filepath.Join(dir, name), os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0o644)
+	})
 
 	log.Debug("yakit directories initialized")
 }
