+++ common/yakgrpc/grpc_language_find_database_test.go
@@ -1,3 +1,5 @@
+//go:build !irify_exclude
+
 package yakgrpc
 
 import (
