+++ common/yakgrpc/grpc_language_suggestion_syntaxflow_stub.go
@@ -1,0 +1,9 @@
+//go:build irify_exclude
+
+package yakgrpc
+
+import "github.com/yaklang/yaklang/common/yakgrpc/ypb"
+
+func SyntaxFlowServer(req *ypb.YaklangLanguageSuggestionRequest) (*ypb.YaklangLanguageSuggestionResponse, bool) {
+	return nil, false
+}
