+++ common/yakgrpc/grpc_smoking_evaluate_sf_rule_test.go
@@ -1,3 +1,5 @@
+//go:build !irify_exclude
+
 package yakgrpc
 
 import (
