+++ common/yakgrpc/grpc_ssa_risk_test.go
@@ -1,3 +1,5 @@
+//go:build !irify_exclude
+
 package yakgrpc
 
 import (
