+++ common/yakgrpc/grpc_syntaxflow_rule_update_test.go
@@ -1,3 +1,5 @@
+//go:build !irify_exclude
+
 package yakgrpc
 
 import (
